DCAC-153: Eliminate duplicate tests.
